refactor move brickSummary to editorNodeLayout
