+++ src/pageEditor/tabs/editTab/editorNodeLayout/brickSummary.module.scss
