+++ src/pageEditor/tabs/editTab/editorNodeLayout/usePipelineNodes.tsx
@@ -80,7 +80,7 @@ import { type Brick } from "@/types/brickTypes";
 import { isNullOrBlank } from "@/utils/stringUtils";
 import { joinName, joinPathParts } from "@/utils/formUtils";
 import { SCROLL_TO_DOCUMENT_PREVIEW_ELEMENT_EVENT } from "@/pageEditor/documentBuilder/preview/ElementPreview";
-import { getBrickOutlineSummary } from "@/pageEditor/brickSummary";
+import { getBrickOutlineSummary } from "@/pageEditor/tabs/editTab/editorNodeLayout/brickSummary";
 import { BrickTypes } from "@/runtime/runtimeTypes";
 
 const ADD_MESSAGE = "Add more bricks with the plus button";
